feature(new integration): Pinterest Conversion API adding event multiplexing (#1428)
* feature(new integration): Pinterest Conversion API adding event multiplexing

* feature(new integration): Pinterest Conversion API review comments addressed

* feature(new integration): Pinterest Conversion API review minor refactor

* Update v0/destinations/pinterest_tag/transform.js


* Update v0/destinations/pinterest_tag/transform.js
