+++ v0/destinations/pinterest_tag/utils.js
@@ -6,9 +6,9 @@ const {
   constructPayload,
   isDefinedAndNotNull,
   isDefined,
-  getHashFromArray
+  getHashFromArrayWithDuplicate
 } = require("../../util");
-const { COMMON_CONFIGS } = require("./config");
+const { COMMON_CONFIGS, CUSTOM_CONFIGS } = require("./config");
 
 const VALID_ACTION_SOURCES = ["app_android", "app_ios", "web", "offline"];
 
@@ -159,11 +159,16 @@ const deduceTrackScreenEventName = (message, Config) => {
           the event mapping in the UI. In case it is similar, will map to that.
    */
   if (Config.eventsMapping.length > 0) {
-    const keyMap = getHashFromArray(Config.eventsMapping, "from", "to", false);
+    const keyMap = getHashFromArrayWithDuplicate(
+      Config.eventsMapping,
+      "from",
+      "to",
+      false
+    );
     eventName = keyMap[trackEventOrScreenName];
   }
   if (isDefined(eventName)) {
-    return eventName;
+    return [...eventName];
   }
 
   /*
@@ -180,14 +185,14 @@ const deduceTrackScreenEventName = (message, Config) => {
     });
 
     if (isDefinedAndNotNull(eventMapInfo)) {
-      return eventMapInfo.dest;
+      return [eventMapInfo.dest];
     }
   }
 
   /*
   Step 3: In case both of the above stated cases fail, will mark the event as "custom"
  */
-  return "custom";
+  return ["custom"];
 };
 
 /**
@@ -202,12 +207,12 @@ const deduceTrackScreenEventName = (message, Config) => {
  */
 const deduceEventName = (message, Config) => {
   const { type } = message;
-  let eventName = "";
+  let eventName = [];
   switch (type) {
     case EventType.PAGE:
       eventName = isDefinedAndNotNull(message.category)
-        ? "ViewCategory"
-        : "PageVisit";
+        ? ["ViewCategory"]
+        : ["PageVisit"];
       break;
     case EventType.TRACK:
     case EventType.SCREEN:
@@ -216,7 +221,6 @@ const deduceEventName = (message, Config) => {
     default:
       throw new CustomError(`The event of type ${type} is not supported`, 400);
   }
-
   return eventName;
 };
 
@@ -283,6 +287,81 @@ const processHashedUserPayload = (userPayload, message) => {
   return processedHashedUserPayload;
 };
 
+/**
+ * This function will process the ecommerce fields and return the final payload
+ * @param {*} message
+ * @param {*} mandatoryPayload
+ * @returns
+ */
+const postProcessEcomFields = (message, mandatoryPayload) => {
+  let totalQuantity = 0;
+  let quantityInconsistent = false;
+  const contentArray = [];
+  const contentIds = [];
+  const { properties } = message;
+  // ref: https://s.pinimg.com/ct/docs/conversions_api/dist/v3.html
+  let customPayload = constructPayload(message, CUSTOM_CONFIGS);
+
+  // if product array is present will look for the product level information
+  if (
+    properties.products &&
+    Array.isArray(properties.products) &&
+    properties.products.length > 0
+  ) {
+    const { products } = properties;
+    products.forEach(product => {
+      const prodParams = setIdPriceQuantity(product, message);
+      contentIds.push(prodParams.contentId);
+      contentArray.push(prodParams.content);
+      if (!product.quantity) {
+        quantityInconsistent = true;
+      }
+      totalQuantity = product.quantity
+        ? totalQuantity + product.quantity
+        : totalQuantity;
+    });
+
+    if (totalQuantity === 0) {
+      /*
+      in case any of the products inside product array does not have quantity,
+       will map the quantity of root level
+      */
+      totalQuantity = properties.quantity;
+    }
+  } else {
+    /*
+    for the events where product array is not present, root level id, price and
+    quantity are taken into consideration
+    */
+    const prodParams = setIdPriceQuantity(message.properties, message);
+    contentIds.push(prodParams.contentId);
+    contentArray.push(prodParams.content);
+    totalQuantity = properties.quantity
+      ? totalQuantity + properties.quantity
+      : totalQuantity;
+  }
+  /*
+    if properties.numOfItems is not provided by the user, the total quantity of the products
+    will be sent as num_items
+  */
+  if (
+    !isDefinedAndNotNull(customPayload.num_items) &&
+    quantityInconsistent === false
+  ) {
+    customPayload.num_items = parseInt(totalQuantity, 10);
+  }
+  customPayload = {
+    ...customPayload,
+    content_ids: contentIds,
+    contents: contentArray
+  };
+
+  return {
+    ...mandatoryPayload,
+    custom_data: { ...customPayload }
+  };
+};
+
 module.exports = {
   processUserPayload,
   processCommonPayload,
@@ -290,5 +369,6 @@ module.exports = {
   setIdPriceQuantity,
   checkUserPayloadValidity,
   processHashedUserPayload,
-  VALID_ACTION_SOURCES
+  VALID_ACTION_SOURCES,
+  postProcessEcomFields
 };
